Refactor/create models folder
