+++ clients/web/src/common/models/Answer.js
@@ -1,4 +1,4 @@
-import { getUniqueId } from './utils'
+import { getUniqueId } from '../utils'
 
 export default class Answer {
   constructor({
