+++ clients/web/src/common/models/Question.js
@@ -1,4 +1,4 @@
-import { getUniqueId } from './utils'
+import { getUniqueId } from '../utils'
 import Section from './Section'
 import Subcategory from './Subcategory'
