+++ clients/web/src/common/models/Subcategory.js
@@ -1,4 +1,4 @@
-import { getUniqueId } from './utils'
+import { getUniqueId } from '../utils'
 
 export default class Subcategory {
   constructor({ id = getUniqueId(), name }) {
